add booking screen ui finished

+++ lib/src/modules/places/data/models/place_edit_form.dart
@@ -15,7 +15,7 @@ class PlaceEditForm {
 
   ///
   int ownerId;
-  int? minimumHours;
+  double? minimumHours;
 
   ///
   List<String> images;
